run http connections in separate IO thread
eventually, we should run the whole http connection in the database session/
thread. however, the current http stack expects to run on an event loop. since
the http code is rubbish and will eventually be replaced by something proper,
we're spwaning a separate IO thread that runs the old event loop code as a
workaround for now.